Mise à jour de la carte
